tidy up JDataClass and JDataClassAuto
