+++ kondor-auto/src/test/kotlin/com/ubertob/kondor/json/JDataClassAutoTest.kt
@@ -1,0 +1,19 @@
+package com.ubertob.kondor.json
+
+import org.junit.jupiter.api.Test
+
+class JDataClassAutoTest {
+
+    object PersonRefl : JDataClassAuto<Person>(Person::class)
+
+    @Test
+    fun `JDataClassAuto doesn't need the fields declaration`() {
+
+        PersonRefl.registerAllProperties() //temp hack
+
+        PersonRefl.toJson(randomPerson(), JsonStyle.prettyWithNulls)
+
+        PersonRefl.testParserAndRender(100) { randomPerson() }
+
+    }
+}
