+++ kondor-auto/src/test/kotlin/com/ubertob/kondor/json/JDataClassTest.kt
@@ -133,16 +133,4 @@ class JDataClassTest {
     }
 
 
-    object PersonRefl : JDataClassAuto<Person>(Person::class)
-
-    @Test
-    fun `JDataClassAuto doesn't need the fields declaration`() {
-
-        PersonRefl.registerAllProperties() //temp hack
-
-        PersonRefl.toJson(randomPerson(), JsonStyle.prettyWithNulls)
-
-        PersonRefl.testParserAndRender(100) { randomPerson() }
-
-    }
 }
